updated findHorizonTime()
Added an API endpoint for finding flight horizon @ http://127.0.0.1:5000/flight_horizon. Updated debug print statement and logic for findHorizonTime(). The findHorizonTime() needs more work, currently as soon horizon is missing either a peak or rise, the function breaks and return whatever it was able to find during the given duration.